+++ src/main/java/net/grid/vampiresdelight/common/utility/VDNameUtils.java
@@ -3,19 +3,20 @@ package net.grid.vampiresdelight.common.utility;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.block.Block;
 
 @SuppressWarnings("unused")
 public class VDNameUtils {
-    public static ResourceLocation itemLocation(Item item) {
-        return BuiltInRegistries.ITEM.getKey(item);
+    public static ResourceLocation itemLocation(ItemLike item) {
+        return BuiltInRegistries.ITEM.getKey(item.asItem());
     }
 
-    public static String itemName(Item item) {
+    public static String itemName(ItemLike item) {
         return itemLocation(item).getPath();
     }
 
-    public static String itemNamespace(Item item) {
+    public static String itemNamespace(ItemLike item) {
         return itemLocation(item).getNamespace();
     }
     public static Item getItem(ResourceLocation key) {